requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic>=2.9.0
requests==2.31.0
duckdb>=1.0.0
psycopg2-binary>=2.9.10
langchain==0.1.0
langchain-community==0.0.19
pydantic-settings>=2.3.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-jaeger==1.21.0
sentence-transformers>=2.2.0
tqdm>=4.66.0
httpx>=0.24.0

